// src/app/models/payment.model.ts
export default class PurchaseModel {
  id: number = 0;
  status: string = "";
  reference: string = "";
  receipt_number: string = "";
  amount: string = '';
  message: string = "";
  gateway_response: string = "";
  paid_at: string = "";
  created_at: string = "";
  channel: string = "";
  currency: string = "";
  ip_address: string = "";
  log: string = "";
  fees: string = "";
  requested_amount: string = "";
  transaction_date: string = "";
  customercode: string = "";
  paidat: string = "";}